usb: phy: Increase size of override PHY init sequence
The current size of runtime override PHY init sequence module
parameter is 20. This is not enough if user wants to input long
PHY init sequence consisting of 7 or more tuples of
(offset, val, delay). Increase this size to 100.

override_phy_init usage:
Enter PHY init sequence in series of tuples (offset, val, delay).
Keep a (-1, -1, 0) to denote the end of the sequence.

 Example to set {(0xc0,0x01,0),(0xe8,0x0d,0)} sequence:

 echo "0xc0,0x01,0,0xe8,0x0d,0,-1,-1,0" >
   /sys/module/phy_qcom_snps_28nm_hs/parameters/override_phy_init

Change-Id: I8e646602f6fa3cc8edd083428436402c29190961
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>